requirements.txt
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0

pip~=21.2.4
py~=1.10.0
pytest~=6.2.5
setuptools~=57.4.0